+++ CHANGELOG.md
@@ -5,10 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Can now increment playCount from start of song as well as from end
+- Can choose start/end at setup time, or while plugin is running via messages
+
 ## [0.0.1] - 2026-05-09
 
 ### Added
-
 - Initial release
 - playCount sticker updated when song is nearly finished
 - Padding (time between update moment and end of song) configurable at setup time
@@ -17,4 +22,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Sticker name configurable at setup time
 - Sticker name configurable by passing messages
 
+[Unreleased]: https://github.com/username/repo/compare/v0.0.1...HEAD
 [0.0.1]: https://github.com/username/repo/releases/tag/v0.0.1